add request_cs control msg to switches, change timestamp to unsigned long

+++ node/NodeNetwork.h
@@ -22,7 +22,7 @@ class NodeNetwork: public SocketEventListener
         int init();
         int start();
 
-        int send(unsigned int from,unsigned int to,unsigned int timestamp, char* message);
+        int send(unsigned int from,unsigned int to,unsigned long timestamp, char* message);
 
         virtual ~NodeNetwork();
